[fix] 修复BigDecimal精度问题

+++ module_common/src/main/kotlin/life/chenshi/keepaccounts/module/common/database/Converters.kt
@@ -9,25 +9,27 @@ import java.util.*
  */
 class Converters {
 
-    // 时间戳转日期
+    // 时间戳转日期 取出
     @TypeConverter
     fun fromTimestamp(value: Long?): Date? {
         return value?.let { Date(it * 1000L) }
     }
 
-    // 日期转时间戳
+    // 日期转时间戳 存入
     @TypeConverter
     fun dateToTimestamp(date: Date?): Long? {
         return date?.time?.div(1000)
     }
 
+    // 取出
     @TypeConverter
-    fun fromInt(value: Int?): BigDecimal? {
+    fun fromInt(value: Long?): BigDecimal? {
         return value?.let { BigDecimal(it).movePointLeft(2) }
     }
 
+    // 存入 sqlite3不支持BigDecimal, 需要手动转换
     @TypeConverter
-    fun bigDecimalToInt(bigDecimal: BigDecimal?): Int? {
-        return bigDecimal?.toInt()
+    fun bigDecimalToInt(bigDecimal: BigDecimal?): Long? {
+        return bigDecimal?.toLong()
     }
 }
